17: refactor state related functions into state.py
